# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Bauk Cheescake",
        "version": "0.0.11",
        "description": "appiid: 3232f21e-653e-4e98-b46d-c34cbc751182 - TODO: All users can post new addresses"
    },
    "servers": [
        {
            "url": "https://localhost:443"
        }
    ],
    "security": [
        {
            "BearerAuth": [],
            "Refresh": [],
            "MailAuth": [],
            "AppiId": []
        }
    ],
    "paths": {
...
        "/products/{id}": {
            "get": {
                "security": [
                    {
                        "AppiId": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "description": "Get products by id",
                "tags": [
                    "Products"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "Application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/products/": {
            "get": {
                "security": [
                    {
                        "AppiId": []
                    }
                ],
                "description": "Get products",
                "tags": [
                    "Products"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "Application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AppiId": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Post products",
                "requestBody": {
                    "content": {
                        "Application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostProduct"
                            }
                        }
                    }
                },
                "tags": [
                    "Products"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorizaiton"
                    }
                }
            }
        },
...
        "/sizes/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "Application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Size"
                            }
                        }
                    }
                },
                "description": "Get products",
                "tags": [
                    "Sizes"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "get": {
                "security": [
                    {
                        "AppiId": []
                    }
                ],
                "description": "Get products",
                "tags": [
                    "Sizes"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/sizes/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "Application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Size"
                            }
                        }
                    }
                },
                "description": "",
                "tags": [
                    "Sizes"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "get": {
                "security": [
                    {
                        "AppiId": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "description": "",
                "tags": [
                    "Sizes"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/sizes/multi": {
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "Application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Size"
                                }
                            }
                        }
                    }
                },
                "description": "Get products",
                "tags": [
                    "Sizes"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "Refresh": {
                "type": "apiKey",
                "in": "header",
                "name": "Refresh"
            },
            "MailAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "MailToken"
            },
            "AppiId": {
                "type": "apiKey",
                "in": "header",
                "name": "appiId"
            }
        },
        "schemas": {
...
            "Product": {
                "type": "object",
                "properties": {
                    "ID": {
                        "type": "integer"
                    },
                    "Name": {
                        "type": "string"
                    },
                    "Available": {
                        "type": "boolean"
                    },
                    "Description": {
                        "type": "string"
                    },
                    "File_Path": {
                        "type": "string"
                    },
                    "Size": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "ID": {
                                    "type": "integer"
                                },
                                "Size": {
                                    "type": "string"
                                },
                                "Price": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "Additions": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "ID": {
                                    "type": "integer"
                                },
                                "Name": {
                                    "type": "string"
                                },
                                "Price": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            },
            "PostProduct": {
                "type": "object",
                "properties": {
                    "IdPicture": {
                        "type": "integer"
                    },
                    "Name": {
                        "type": "string"
                    },
                    "Description": {
                        "type": "string"
                    },
                    "Available": {
                        "type": "boolean"
                    },
                    "Sizes": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "IdSize": {
                                    "type": "integer"
                                },
                                "Price": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "Additions": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "IdAddition": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            },
            "Size": {
                "type": "object",
                "properties": {
                    "Size": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}